+++ includes/class-mjb-feeds.php
@@ -9,6 +9,8 @@ if (!defined('ABSPATH')) {
 
 class MJB_Feeds
 {
+    const FEED_LIMIT = 100;
+
     /**
      * Initialize Feeds.
      */
@@ -27,6 +29,18 @@ class MJB_Feeds
     }
 
     /**
+     * Build feed query args.
+     *
+     * @return array
+     */
+    public static function build_feed_query_args()
+    {
+        return MJB_Search::build_query_args(array(), array(
+            'posts_per_page' => self::FEED_LIMIT,
+        ));
+    }
+
+    /**
      * Render Job Feed (XML).
      */
     public function render_job_feed()
@@ -34,10 +48,11 @@ class MJB_Feeds
         header('Content-Type: application/xml; charset=UTF-8');
         echo '<?xml version="1.0" encoding="UTF-8"?>';
         ?>
-        <rss version="2.0" xmlns:content="http://purl.org/rss/1.0/modules/content/"
-            xmlns:wfw="http://wellformedweb.org/CommentAPI/" xmlns:dc="http://purl.org/dc/elements/1.1/"
-            xmlns:atom="http://www.w3.org/2005/Atom" xmlns:sy="http://purl.org/rss/1.0/modules/syndication/"
-            xmlns:slash="http://purl.org/rss/1.0/modules/slash/">
+        <rss version="2.0"
+            xmlns:content="http://purl.org/rss/1.0/modules/content/"
+            xmlns:dc="http://purl.org/dc/elements/1.1/"
+            xmlns:atom="http://www.w3.org/2005/Atom"
+            xmlns:mjb="https://github.com/MartinOrton/modern-job-board/ns/feed/1.0">
             <channel>
                 <title><?php echo get_bloginfo_rss('name'); ?> - Job Listings</title>
                 <atom:link href="<?php self_link(); ?>" rel="self" type="application/rss+xml" />
@@ -47,32 +62,33 @@ class MJB_Feeds
                 </lastBuildDate>
                 <language><?php bloginfo_rss('language'); ?></language>
                 <?php
-                $args = array(
-                    'post_type' => 'job_listing',
-                    'post_status' => 'publish',
-                    'posts_per_page' => 100, // Limit for feed
-                );
-                $query = new WP_Query($args);
+                $query = new WP_Query(self::build_feed_query_args());
 
-                if ($query->have_posts()):
-                    while ($query->have_posts()):
+                if ($query->have_posts()) :
+                    while ($query->have_posts()) :
                         $query->the_post();
-                        $company_name = get_post_meta(get_the_ID(), '_company_name', true);
-                        $location = get_the_term_list(get_the_ID(), 'job_location', '', ', ');
-                        $type = get_the_term_list(get_the_ID(), 'job_type', '', ', ');
+                        $post_id = get_the_ID();
+                        $company_name = get_post_meta($post_id, '_company_name', true);
+                        $location_terms = wp_get_post_terms($post_id, 'job_location', array('fields' => 'names'));
+                        $type_terms = wp_get_post_terms($post_id, 'job_type', array('fields' => 'names'));
+                        $location = !empty($location_terms) ? $location_terms[0] : '';
+                        $type = !empty($type_terms) ? $type_terms[0] : '';
+                        $application_url = get_permalink($post_id);
                         ?>
                         <item>
                             <title><?php the_title_rss(); ?></title>
                             <link><?php the_permalink_rss(); ?></link>
-                            <guid isPermaLink="false"><?php the_guid(); ?></guid>
+                            <guid isPermaLink="true"><?php the_permalink_rss(); ?></guid>
                             <pubDate><?php echo mysql2date('D, d M Y H:i:s +0000', get_post_time('Y-m-d H:i:s', true), false); ?>
                             </pubDate>
                             <dc:creator><![CDATA[<?php the_author(); ?>]]></dc:creator>
                             <description><![CDATA[<?php the_excerpt_rss(); ?>]]></description>
                             <content:encoded><![CDATA[<?php the_content_feed('rss2'); ?>]]></content:encoded>
-                            <jobbranch:company><![CDATA[<?php echo esc_html($company_name); ?>]]></jobbranch:company>
-                            <jobbranch:location><![CDATA[<?php echo strip_tags($location); ?>]]></jobbranch:location>
-                            <jobbranch:type><![CDATA[<?php echo strip_tags($type); ?>]]></jobbranch:type>
+                            <mjb:company><![CDATA[<?php echo esc_html($company_name); ?>]]></mjb:company>
+                            <mjb:location><![CDATA[<?php echo esc_html($location); ?>]]></mjb:location>
+                            <mjb:jobType><![CDATA[<?php echo esc_html($type); ?>]]></mjb:jobType>
+                            <mjb:applyUrl><![CDATA[<?php echo esc_url($application_url); ?>]]></mjb:applyUrl>
+                            <mjb:featured><?php echo get_post_meta($post_id, '_featured', true) ? '1' : '0'; ?></mjb:featured>
                         </item>
                         <?php
                     endwhile;